+++ src/page-support/exclusive-content/components/dependent-script.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import SyntaxHighlighter from 'react-syntax-highlighter'
 import { docco } from 'react-syntax-highlighter/dist/esm/styles/hljs'
+import { CopyContent } from './copy-content'
 
 import { useExclusiveContent } from '../state'
 
@@ -30,9 +31,15 @@ export function DependentScript() {
 
   return (
     <>
-      <SyntaxHighlighter language='html' style={docco}>
-        {dependentScript}
-      </SyntaxHighlighter>
+      <div>
+        <CopyContent
+          id='dependentScript'
+          message='Copy script with dependencies'
+        />
+        <SyntaxHighlighter id='dependentScript' language='html' style={docco}>
+          {dependentScript}
+        </SyntaxHighlighter>
+      </div>
     </>
   )
 }
